pyproject.toml
[project]
name = "psy-bench"
version = "0.1.0"
description = "A library for benchmarking AI model responses to mental health scenarios"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "aiohttp>=3.12.15",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["psy_bench"]
